requirements.txt
fastapi
uvicorn[standard]
sounddevice
soundcard
PyAudioWPatch
numpy
faster-whisper
pydantic
python-dotenv
jinja2
aiofiles
pytest
black
ruff
